set search query now tokenizes query into state
